backend/requirements.txt
Django==5.2.0
djangorestframework==3.15.0
django-cors-headers==4.3.1
Pillow==10.2.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.1
drf-yasg==1.21.7
gunicorn==21.2.0